docker: Re-add xz-utils (#39553)
Previously this was installed as a dependency of mailcap, which was a
dependency of mime-support, which was a dependency of various things.
But apparently things depending on mime-support are transitioning to
media-types which doesn't have that dependency.

We need xz-utils for extracting `build.tar.xz` in CI, so make sure it's
installed.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/11060062524

Upstream-Ref: Automattic/jetpack@14207d752b295ea1293763894645992b9eb54286